+++ sdk/python/pulumi_postgresql/default_privileg.py
@@ -25,6 +25,7 @@ class DefaultPrivileg(pulumi.CustomResource):
                  privileges: Optional[pulumi.Input[Sequence[pulumi.Input[str]]]] = None,
                  role: Optional[pulumi.Input[str]] = None,
                  schema: Optional[pulumi.Input[str]] = None,
+                 with_grant_option: Optional[pulumi.Input[bool]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
@@ -38,6 +39,7 @@ class DefaultPrivileg(pulumi.CustomResource):
         :param pulumi.Input[Sequence[pulumi.Input[str]]] privileges: The list of privileges to apply as default privileges
         :param pulumi.Input[str] role: The name of the role to which grant default privileges on
         :param pulumi.Input[str] schema: The database schema to set default privileges for this role
+        :param pulumi.Input[bool] with_grant_option: Permit the grant recipient to grant it to others
         """
         pulumi.log.warn("""DefaultPrivileg is deprecated: postgresql.DefaultPrivileg has been deprecated in favor of postgresql.DefaultPrivileges""")
         if __name__ is not None:
@@ -72,9 +74,8 @@ class DefaultPrivileg(pulumi.CustomResource):
             if role is None and not opts.urn:
                 raise TypeError("Missing required property 'role'")
             __props__['role'] = role
-            if schema is None and not opts.urn:
-                raise TypeError("Missing required property 'schema'")
             __props__['schema'] = schema
+            __props__['with_grant_option'] = with_grant_option
         super(DefaultPrivileg, __self__).__init__(
             'postgresql:index/defaultPrivileg:DefaultPrivileg',
             resource_name,
@@ -90,7 +91,8 @@ class DefaultPrivileg(pulumi.CustomResource):
             owner: Optional[pulumi.Input[str]] = None,
             privileges: Optional[pulumi.Input[Sequence[pulumi.Input[str]]]] = None,
             role: Optional[pulumi.Input[str]] = None,
-            schema: Optional[pulumi.Input[str]] = None) -> 'DefaultPrivileg':
+            schema: Optional[pulumi.Input[str]] = None,
+            with_grant_option: Optional[pulumi.Input[bool]] = None) -> 'DefaultPrivileg':
         """
         Get an existing DefaultPrivileg resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
@@ -104,6 +106,7 @@ class DefaultPrivileg(pulumi.CustomResource):
         :param pulumi.Input[Sequence[pulumi.Input[str]]] privileges: The list of privileges to apply as default privileges
         :param pulumi.Input[str] role: The name of the role to which grant default privileges on
         :param pulumi.Input[str] schema: The database schema to set default privileges for this role
+        :param pulumi.Input[bool] with_grant_option: Permit the grant recipient to grant it to others
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -115,6 +118,7 @@ class DefaultPrivileg(pulumi.CustomResource):
         __props__["privileges"] = privileges
         __props__["role"] = role
         __props__["schema"] = schema
+        __props__["with_grant_option"] = with_grant_option
         return DefaultPrivileg(resource_name, opts=opts, __props__=__props__)
 
     @property
@@ -159,11 +163,19 @@ class DefaultPrivileg(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def schema(self) -> pulumi.Output[str]:
+    def schema(self) -> pulumi.Output[Optional[str]]:
         """
         The database schema to set default privileges for this role
         """
         return pulumi.get(self, "schema")
+
+    @property
+    @pulumi.getter(name="withGrantOption")
+    def with_grant_option(self) -> pulumi.Output[Optional[bool]]:
+        """
+        Permit the grant recipient to grant it to others
+        """
+        return pulumi.get(self, "with_grant_option")
 
     def translate_output_property(self, prop):
         return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop